requirements.txt
# Core
gradio>=4.0.0
llama-cpp-python>=0.2.0

# News + parsing
requests>=2.28
beautifulsoup4>=4.12
feedparser>=6.0
newspaper4k>=0.9.3
trafilatura>=2.0.0
pykakasi>=2.2.1

# Development helpers (optional)
black
ruff
